frontend/requirements.txt
streamlit
requests
pydantic==2.9.2
python-dotenv
